requirements.txt
alembic==1.17.0
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pypdfium2==5.1.0
python-dotenv==1.1.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.3
